REF: code cleanup

+++ app.py
@@ -36,13 +36,10 @@ def index():
 def get_workout():
     """
     Display workout html page. 
-    Returns:  list of user workouts
+    :return  list of user workouts
     """
     if "user" in session:
         workout_list = list(mongo.db.routines.find().sort("routine_name", 1))
-        # categories = list(mongo.db.categories.find().sort("category_name", 1))
-        # return render_template("categories.html", categories=categories)
-
         return render_template("workout.html", workout_list=workout_list)
     else:
         flash("You need to be logged in to view workouts")
@@ -56,20 +53,9 @@ def create_workout():
     :return: [status]
     """
     if "user" in session:
-        # exercise_list_user = list(mongo.db.exercises.find({"created_by": {"$eq": session["user"]}}))
-        # exercise_list_admin = list(mongo.db.exercises.find({"created_by": {"$eq": "admin"}}))
-
-        # for i in range(len(exercise_list_user)):
-        #     for b in range(len(exercise_list_admin)):
-        #         if exercise_list_user[i]["exercise_name"] in exercise_list_admin[b]["exercise_name"]:
-        #             exercise_list_admin.pop(b)
-        # exercise_list = exercise_list_user + exercise_list_admin
-
         exercise_list = list(mongo.db.exercises.find({"$or": [{"created_by": {"$eq": "admin"}},
                                                               {"created_by": {"$eq": session["user"]}}]}
-
                                                      ))
-
         if request.method == "POST":
             if request.method == "POST":
                 submit = {
@@ -84,8 +70,6 @@ def create_workout():
                     'saved': False,
                     "created_by": session["user"]
                 }
-
-                # print("submit", submit)
                 mongo.db.routines.insert_one(submit)
                 flash_text = "{} Successfully Created".format(submit["workout_name"])
                 flash(flash_text)
@@ -119,30 +103,17 @@ def edit_workout(workout_id):
                 'saved': is_saved,
                 "created_by": session["user"]
             }
-
             mongo.db.routines.update({"_id": ObjectId(workout_id)}, submit)
             flash_text = "{} Successfully Updated".format(submit["workout_name"])
             flash(flash_text)
             print(flash_text)
             return redirect(url_for("get_workout"))
 
-        # print(request.form)
         single_workout = mongo.db.routines.find_one({"_id": ObjectId(workout_id)})
-
-        # exercise_list_user = list(mongo.db.exercises.find({"created_by": {"$eq": session["user"]}}))
-        # exercise_list_admin = list(mongo.db.exercises.find({"created_by": {"$eq": "admin"}}))
-        #
-        # # for i in range(len(exercise_list_user)):
-        # #     for b in range(len(exercise_list_admin)):
-        # #         if exercise_list_user[i]["exercise_name"] in exercise_list_admin[b]["exercise_name"]:
-        # #             exercise_list_admin.pop(b)
-        # exercise_list = exercise_list_user + exercise_list_admin
-
         exercise_list = list(mongo.db.exercises.find({"$or": [{"created_by": {"$eq": "admin"}},
                                                               {"created_by": {"$eq": session["user"]}}]}
 
                                                      ))
-
         return render_template("edit_workout.html",
                                workout_list=single_workout,
                                exercise_list=exercise_list)
@@ -168,7 +139,6 @@ def start_workout(workout_id):
     Start existing workout. Takes arguments: [workout_id], query DB,
         :return workout_list, exercise_history, exercise_list
     """
-
     if "user" in session:
         single_workout = mongo.db.routines.find_one({"_id": ObjectId(workout_id)})
         exercise_list = list(mongo.db.exercises.find())
@@ -209,7 +179,6 @@ def update_workout_data(query):
                 }
             mongo.db.user_profile.insert_one(user_profile_data)
 
-        # print(exercise_history)
         submit = {
             "exercise_name": exercise_data["exercise_name"],
             "description": exercise_data["description"],
@@ -224,10 +193,9 @@ def update_workout_data(query):
             'yt_url': exercise_data["yt_url"],
             'steps': exercise_data["steps"],
             "created_by": exercise_data["created_by"],
-            "exercise_history": exercise_history  # request.form.getlist("info_"+query+"[]")
+            "exercise_history": exercise_history
         }
-        # print(submit['exercise_history'])
-        # mongo.db.user_profile.insert_one(user_profile_data)
+
         mongo.db.exercises.update({"_id": ObjectId(query)}, submit)
 
         return redirect(url_for("start_workout", workout_id=workout_id))
@@ -242,14 +210,8 @@ def get_exercise_list():
         Return: exercise_list, exercise_list_admin
     """
     if "user" in session:
-
-        user = list(mongo.db.exercises.find({
-            "$and": [{"created_by": {'$eq': session["user"]}}]
-        }))
-        admin = list(mongo.db.exercises.find({
-            "$and": [{"created_by": {'$eq': "admin"}}]
-        }))
-        # exercise_list = list(mongo.db.exercises.find().sort("exercise_name", 1))
+        user = list(mongo.db.exercises.find({"$and": [{"created_by": {'$eq': session["user"]}}]}))
+        admin = list(mongo.db.exercises.find({"$and": [{"created_by": {'$eq': "admin"}}]}))
         return render_template("exercise_all.html",
                                exercise_list=user,
                                exercise_list_admin=admin)
@@ -262,7 +224,6 @@ def get_exercise_list():
 def create_exercise():
     """
     Create new exercise, read form data validate youtube url, check if image provided before store to DB
-
     """
     if "user" in session:
         pattern = re.compile("(http(s?):)([/|.|\w|\s|-])*\.(?:jpg|gif|png|jpeg)")
@@ -303,9 +264,7 @@ def create_exercise():
                 "is_system": False,
                 "exercise_history": []
             }
-
             mongo.db.exercises.insert_one(submit)
-            # print(submit)
             flash_text = "{} Successfully Created".format(submit["exercise_name"])
             flash(flash_text)
             print(flash_text)
@@ -385,7 +344,6 @@ def edit_exercise(exercise_id):
         return render_template("exercise_edit_single.html",
                                exercise=single_exercise,
                                exercise_category_list=exercise_category_list)
-
     else:
         flash("You need to be logged in to perform this action")
         return redirect(url_for("login"))
@@ -393,14 +351,18 @@ def edit_exercise(exercise_id):
 
 @app.route("/exercise/delete/<exercise_id>", methods=["GET", "POST"])
 def delete_exercise(exercise_id):
-    # mongo.db.exercises.remove({"_id": ObjectId(exercise_id)})
+    """
+    Search user_profile DB for exercise history and delete any record containing exercise_id
+    :param exercise_id:
+    :return: message to user
+    """
     if "user" in session:
-
         get_exercise_history = mongo.db.user_profile.find(
             {"exercise_id": {"$eq": exercise_id}}
         )
-        for exercise in get_exercise_history:
-            mongo.db.user_profile.delete_many({"_id": ObjectId(exercise['_id'])})
+        if get_exercise_history:
+            for exercise in get_exercise_history:
+                mongo.db.user_profile.delete_many({"_id": ObjectId(exercise['_id'])})
         mongo.db.exercises.delete_many({"_id": ObjectId(exercise_id)})
 
         flash("Exercise Successfully Deleted")
@@ -421,11 +383,9 @@ def search():
             exercises = list(mongo.db.exercises.find({"$text": {"$search": query},
                                                       "$and": [{"created_by": {'$eq': session['user']}}]
                                                       }))
-
             admin = list(mongo.db.exercises.find({"$text": {"$search": query},
                                                   "$and": [{"created_by": {'$eq': 'admin'}}]
                                                   }))
-
             return render_template("exercise_all.html", exercise_list=exercises, exercise_list_admin=admin)
         else:
             return redirect(url_for('get_exercise_list'))
@@ -433,10 +393,6 @@ def search():
         flash("You need to be logged in to perform this action")
         return redirect(url_for("login"))
 
-
-# return render_template("exercise_all.html",
-#                        exercise_list=user,
-#                        exercise_list_admin=admin)
 
 # ========================handle login logout register=================================================
 
@@ -447,17 +403,14 @@ def register():
         # check if username already exists in db
         existing_user = mongo.db.users.find_one(
             {"username": request.form.get("username").lower()})
-
         if existing_user:
             flash("Username already exists")
             return redirect(url_for("register"))
-
-        register = {
+        register_user = {
             "username": request.form.get("username").lower(),
             "password": generate_password_hash(request.form.get("password"))
         }
-        mongo.db.users.insert_one(register)
-
+        mongo.db.users.insert_one(register_user)
         # put the new user into 'session' cookie
         session["user"] = request.form.get("username").lower()
         flash("Registration Successful!")
@@ -472,7 +425,6 @@ def login():
         # check if username exists in db
         existing_user = mongo.db.users.find_one(
             {"username": request.form.get("username").lower()})
-
         if existing_user:
             # ensure hashed password matches user input
             if check_password_hash(
@@ -486,7 +438,6 @@ def login():
                 # invalid password match
                 flash("Incorrect Username and/or Password")
                 return redirect(url_for("login"))
-
         else:
             # username doesn't exist
             flash("Incorrect Username and/or Password")
@@ -509,10 +460,8 @@ def profile():
         })["username"]
     except (TypeError, KeyError):
         return redirect(url_for("login"))
-
     if session["user"]:
         return render_template("profile.html")
-
     return redirect(url_for("login"))
